backend/requirements.txt
# FastAPI and related dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0

# Data processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
statsmodels==0.14.0
joblib==1.3.2

# Optimization
cvxpy==1.4.1

# HTTP client
httpx==0.25.2

# Environment management
python-decouple==3.8

# Caching
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0